requirements.txt
asgiref==3.8.1
Django==5.1.4
django-simple-history==3.7.0
django-widget-tweaks==1.5.0
pillow==11.0.0
sqlparse==0.5.2
tzdata==2024.2
